Removed urlEncode and urlDecode from pre-release. We don't know if this implementation is subject to copyright, so need to replace with a clean implementation (volunteers welcome).
Changed "url" to "key" in many method calls. This is part of an (incomplete) refactor
so that HTTP POST services can be cached. The idea is: you can have a unique hash
key based on additional line(s) of text in the "url" and thus make the service cachable,
this url + '´\n' + uniquePostIdentifier format is referred to as a "key" since it is
a hash key, not just a url.