Added TantalumBlackberry directory to mirror new package scheme
This is a first step to merging in adiksonline's Blackberry port. The
build structure and other aspects of Tantalum have evolved since the port
was forked 3 months ago. The following steps will be to copy by hand key
parts of the TantalumLibrary directory structure relating to BB into the
TantalumBlackberry directory, update them as needed to reflect lib changes
or caching. The new RMSFastCache in the JME port (the TantalumBlackberry
directory is currently just a copy of the TantalumJME directory with minor
build changes) should also work on
Blackberry, perhaps unchanged if luck and design and standards hold.