Refactor to move accessOrder LRUVector from StaticCache level to FlashCache level
We want the accessOrder to contain only items actually in flash memory. We loosly couple the delete process between heap level and flash level (which is synchronized only after slow tasks complete) by asking the flash level for a list of digests that would add up to as much space as we want to clear from the cache. We then delete those object from the help level and check how much space we actually free in the process. This is more robust against concurrent updates such as queued writes, or cases where heap may contain something not yet in flash (we try to delete it, but it isn't actually stored yet so we can't delete it).